fix: resolve test command glob pattern issue by renaming test files to .js
Previously, the glob pattern `src/**/*.test.ts` was not expanding correctly in
the shell, causing only 2 out of 17 test files to be executed. This fix renames
all test files from .ts to .js and updates the test command to use the .js
pattern with proper quotes to ensure all tests are discovered and run.

Fixes #97
