+++ src/cli/handlers/attach.test.js
@@ -1,0 +1,156 @@
+import { deepStrictEqual } from "node:assert";
+import { describe, it, mock } from "node:test";
+import { err, ok } from "../../core/types/result.ts";
+import {
+  BranchNotFoundError,
+  WorktreeAlreadyExistsError,
+} from "../../core/worktree/errors.ts";
+
+const exitWithErrorMock = mock.fn((message, code) => {
+  throw new Error(`Exit with code ${code}: ${message}`);
+});
+const outputLogMock = mock.fn();
+const getGitRootMock = mock.fn();
+const attachWorktreeCoreMock = mock.fn();
+const shellInWorktreeMock = mock.fn();
+const execInWorktreeMock = mock.fn();
+
+mock.module("../errors.ts", {
+  namedExports: {
+    exitWithError: exitWithErrorMock,
+    exitCodes: {
+      validationError: 3,
+      notFound: 2,
+      generalError: 1,
+      success: 0,
+    },
+  },
+});
+
+mock.module("../output.ts", {
+  namedExports: {
+    output: { log: outputLogMock },
+  },
+});
+
+mock.module("../../core/git/libs/get-git-root.ts", {
+  namedExports: {
+    getGitRoot: getGitRootMock,
+  },
+});
+
+mock.module("../../core/worktree/attach.ts", {
+  namedExports: {
+    attachWorktreeCore: attachWorktreeCoreMock,
+  },
+});
+
+mock.module("../../core/process/shell.ts", {
+  namedExports: {
+    shellInWorktree: shellInWorktreeMock,
+  },
+});
+
+mock.module("../../core/process/exec.ts", {
+  namedExports: {
+    execInWorktree: execInWorktreeMock,
+  },
+});
+
+const { attachHandler } = await import("./attach.ts");
+
+describe("attachHandler", () => {
+  it("should attach to existing branch successfully", async () => {
+    exitWithErrorMock.mock.resetCalls();
+    outputLogMock.mock.resetCalls();
+    getGitRootMock.mock.mockImplementation(() => Promise.resolve("/repo"));
+    attachWorktreeCoreMock.mock.mockImplementation(() =>
+      Promise.resolve(ok("/repo/.git/phantom/worktrees/feature")),
+    );
+
+    await attachHandler(["feature"]);
+
+    deepStrictEqual(exitWithErrorMock.mock.calls.length, 0);
+    deepStrictEqual(
+      outputLogMock.mock.calls[0].arguments[0],
+      "Attached phantom: feature",
+    );
+    deepStrictEqual(attachWorktreeCoreMock.mock.calls[0].arguments, [
+      "/repo",
+      "feature",
+    ]);
+  });
+
+  it("should exit with error when no branch name provided", async () => {
+    exitWithErrorMock.mock.resetCalls();
+
+    try {
+      await attachHandler([]);
+    } catch (error) {
+      // Expected to throw
+    }
+
+    deepStrictEqual(exitWithErrorMock.mock.calls[0].arguments, [
+      "Missing required argument: branch name",
+      3,
+    ]);
+  });
+
+  it("should exit with error when both --shell and --exec are provided", async () => {
+    exitWithErrorMock.mock.resetCalls();
+    getGitRootMock.mock.resetCalls();
+    attachWorktreeCoreMock.mock.resetCalls();
+
+    try {
+      await attachHandler(["feature", "--shell", "--exec", "ls"]);
+    } catch (error) {
+      // Expected to throw
+    }
+
+    deepStrictEqual(exitWithErrorMock.mock.calls[0].arguments, [
+      "Cannot use both --shell and --exec options",
+      3,
+    ]);
+    deepStrictEqual(getGitRootMock.mock.calls.length, 0);
+    deepStrictEqual(attachWorktreeCoreMock.mock.calls.length, 0);
+  });
+
+  it("should handle BranchNotFoundError", async () => {
+    exitWithErrorMock.mock.resetCalls();
+    getGitRootMock.mock.mockImplementation(() => Promise.resolve("/repo"));
+    attachWorktreeCoreMock.mock.mockImplementation(() =>
+      Promise.resolve(err(new BranchNotFoundError("nonexistent"))),
+    );
+
+    try {
+      await attachHandler(["nonexistent"]);
+    } catch (error) {
+      // Expected to throw
+    }
+
+    deepStrictEqual(exitWithErrorMock.mock.calls[0].arguments, [
+      "Branch 'nonexistent' not found",
+      2,
+    ]);
+  });
+
+  it("should spawn shell when --shell flag is provided", async () => {
+    exitWithErrorMock.mock.resetCalls();
+    outputLogMock.mock.resetCalls();
+    shellInWorktreeMock.mock.resetCalls();
+    shellInWorktreeMock.mock.mockImplementation(() =>
+      Promise.resolve(ok({ exitCode: 0 })),
+    );
+    getGitRootMock.mock.mockImplementation(() => Promise.resolve("/repo"));
+    attachWorktreeCoreMock.mock.mockImplementation(() =>
+      Promise.resolve(ok("/repo/.git/phantom/worktrees/feature")),
+    );
+
+    await attachHandler(["feature", "--shell"]);
+
+    deepStrictEqual(shellInWorktreeMock.mock.calls[0].arguments, [
+      "/repo",
+      "feature",
+    ]);
+  });
+});
