+++ src/core/paths.test.js
