+++ src/core/process/shell.test.js
@@ -1,49 +1,52 @@
 import { deepStrictEqual, strictEqual } from "node:assert";
-import { beforeEach, describe, it, mock } from "node:test";
+import { describe, it, mock } from "node:test";
 import { isErr, isOk } from "../types/result.ts";
 import { WorktreeNotFoundError } from "../worktree/errors.ts";
 import { ProcessSpawnError } from "./errors.ts";
-import type { SpawnConfig } from "./spawn.ts";
+// import type { SpawnConfig } from "./spawn.ts";
+
+const validateMock = mock.fn();
+const spawnMock = mock.fn();
+
+mock.module("../worktree/validate.ts", {
+  namedExports: {
+    validateWorktreeExists: validateMock,
+  },
+});
+
+mock.module("./spawn.ts", {
+  namedExports: {
+    spawnProcess: spawnMock,
+  },
+});
+
+const { shellInWorktree } = await import("./shell.ts");
 
 describe("shellInWorktree", () => {
-  let validateMock: ReturnType<typeof mock.fn>;
-  let spawnMock: ReturnType<typeof mock.fn>;
-  let originalShell: string | undefined;
+  let originalShell;
 
-  beforeEach(() => {
+  const resetMocks = () => {
+    validateMock.mock.resetCalls();
+    spawnMock.mock.resetCalls();
     originalShell = process.env.SHELL;
-  });
+  };
 
   it("should spawn shell successfully when worktree exists", async () => {
+    resetMocks();
     process.env.SHELL = "/bin/bash";
-
-    validateMock = mock.fn(() =>
+    validateMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/my-feature",
       }),
     );
-
-    spawnMock = mock.fn(() =>
+    spawnMock.mock.mockImplementation(() =>
       Promise.resolve({
         ok: true,
         value: { exitCode: 0 },
       }),
     );
 
-    mock.module("../worktree/validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("./spawn.ts", {
-      namedExports: {
-        spawnProcess: spawnMock,
-      },
-    });
-
-    const { shellInWorktree } = await import("./shell.ts");
     const result = await shellInWorktree("/test/repo", "my-feature");
 
     strictEqual(isOk(result), true);
@@ -51,14 +54,14 @@ describe("shellInWorktree", () => {
       deepStrictEqual(result.value, { exitCode: 0 });
     }
 
-    const spawnCall = spawnMock.mock.calls[0].arguments[0] as SpawnConfig;
+    const spawnCall = spawnMock.mock.calls[0].arguments[0];
     deepStrictEqual(spawnCall.command, "/bin/bash");
     deepStrictEqual(spawnCall.args, []);
     deepStrictEqual(
       spawnCall.options?.cwd,
       "/test/repo/.git/phantom/worktrees/my-feature",
     );
-    const env = spawnCall.options?.env as NodeJS.ProcessEnv;
+    const env = spawnCall.options?.env;
     deepStrictEqual(env.PHANTOM, "1");
     deepStrictEqual(env.PHANTOM_NAME, "my-feature");
     deepStrictEqual(
@@ -66,70 +69,51 @@ describe("shellInWorktree", () => {
       "/test/repo/.git/phantom/worktrees/my-feature",
     );
 
-    process.env.SHELL = originalShell;
+    // Restore original shell
+    if (originalShell !== undefined) {
+      process.env.SHELL = originalShell;
+    } else {
+      Reflect.deleteProperty(process.env, "SHELL");
+    }
   });
 
   it("should use /bin/sh when SHELL env var is not set", async () => {
-    process.env.SHELL = undefined;
-
-    validateMock = mock.fn(() =>
+    resetMocks();
+    Reflect.deleteProperty(process.env, "SHELL");
+    validateMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
-
-    spawnMock = mock.fn(() =>
+    spawnMock.mock.mockImplementation(() =>
       Promise.resolve({
         ok: true,
         value: { exitCode: 0 },
       }),
     );
 
-    mock.module("../worktree/validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("./spawn.ts", {
-      namedExports: {
-        spawnProcess: spawnMock,
-      },
-    });
-
-    const { shellInWorktree } = await import("./shell.ts");
     await shellInWorktree("/test/repo", "feature");
 
-    deepStrictEqual(
-      (spawnMock.mock.calls[0].arguments[0] as SpawnConfig).command,
-      "/bin/sh",
-    );
+    deepStrictEqual(spawnMock.mock.calls[0].arguments[0].command, "/bin/sh");
 
-    process.env.SHELL = originalShell;
+    // Restore original shell
+    if (originalShell !== undefined) {
+      process.env.SHELL = originalShell;
+    } else {
+      Reflect.deleteProperty(process.env, "SHELL");
+    }
   });
 
   it("should return error when worktree does not exist", async () => {
-    validateMock = mock.fn(() =>
+    resetMocks();
+    validateMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: false,
         message: "Worktree 'non-existent' not found",
       }),
     );
 
-    mock.module("../worktree/validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("./spawn.ts", {
-      namedExports: {
-        spawnProcess: spawnMock,
-      },
-    });
-
-    const { shellInWorktree } = await import("./shell.ts");
     const result = await shellInWorktree("/test/repo", "non-existent");
 
     strictEqual(isErr(result), true);
@@ -142,33 +126,20 @@ describe("shellInWorktree", () => {
   });
 
   it("should pass through spawn process errors", async () => {
-    validateMock = mock.fn(() =>
+    resetMocks();
+    validateMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
-
-    spawnMock = mock.fn(() =>
+    spawnMock.mock.mockImplementation(() =>
       Promise.resolve({
         ok: false,
         error: new ProcessSpawnError("/bin/sh", "Shell not found"),
       }),
     );
 
-    mock.module("../worktree/validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("./spawn.ts", {
-      namedExports: {
-        spawnProcess: spawnMock,
-      },
-    });
-
-    const { shellInWorktree } = await import("./shell.ts");
     const result = await shellInWorktree("/test/repo", "feature");
 
     strictEqual(isErr(result), true);
