+++ src/core/process/tmux.test.js
