+++ src/core/types/result.test.js
