+++ src/core/version.test.js
