+++ src/core/worktree/delete.test.js
@@ -7,42 +7,54 @@ import {
   WorktreeNotFoundError,
 } from "./errors.ts";
 
+const validateWorktreeExistsMock = mock.fn();
+const executeGitCommandMock = mock.fn();
+const executeGitCommandInDirectoryMock = mock.fn();
+
+mock.module("./validate.ts", {
+  namedExports: {
+    validateWorktreeExists: validateWorktreeExistsMock,
+  },
+});
+
+mock.module("../git/executor.ts", {
+  namedExports: {
+    executeGitCommand: executeGitCommandMock,
+    executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
+  },
+});
+
+const { deleteWorktree, getWorktreeStatus, removeWorktree, deleteBranch } =
+  await import("./delete.ts");
+
 describe("deleteWorktree", () => {
+  const resetMocks = () => {
+    validateWorktreeExistsMock.mock.resetCalls();
+    executeGitCommandMock.mock.resetCalls();
+    executeGitCommandInDirectoryMock.mock.resetCalls();
+  };
+
   it("should delete worktree and report when branch deletion fails", async () => {
-    const validateMock = mock.fn(() =>
+    resetMocks();
+    validateWorktreeExistsMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
 
-    const executeGitCommandMock = mock.fn((command: string) => {
-      if (command.includes("worktree remove")) {
+    executeGitCommandMock.mock.mockImplementation((command) => {
+      if (command[0] === "worktree" && command[1] === "remove") {
         return Promise.resolve({ stdout: "", stderr: "" });
       }
-      if (command.includes("branch -D")) {
+      if (command[0] === "branch" && command[1] === "-D") {
         return Promise.reject(new Error("error: branch 'feature' not found."));
       }
       return Promise.reject(new Error("Unexpected command"));
     });
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({ stdout: "", stderr: "" }),
     );
-
-    mock.module("./validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { deleteWorktree } = await import("./delete.ts");
 
     const result = await deleteWorktree("/test/repo", "feature");
 
@@ -58,40 +70,26 @@ describe("deleteWorktree", () => {
   });
 
   it("should delete a worktree successfully when no uncommitted changes", async () => {
-    const validateMock = mock.fn(() =>
+    resetMocks();
+    validateWorktreeExistsMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
 
-    const executeGitCommandMock = mock.fn((command: string) => {
-      if (command.includes("worktree remove")) {
+    executeGitCommandMock.mock.mockImplementation((command) => {
+      if (command[0] === "worktree" && command[1] === "remove") {
         return Promise.resolve({ stdout: "", stderr: "" });
       }
-      if (command.includes("branch -D")) {
+      if (command[0] === "branch" && command[1] === "-D") {
         return Promise.resolve({ stdout: "", stderr: "" });
       }
       return Promise.reject(new Error("Unexpected command"));
     });
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({ stdout: "", stderr: "" }),
     );
-
-    mock.module("./validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { deleteWorktree } = await import("./delete.ts");
 
     const result = await deleteWorktree("/test/repo", "feature");
 
@@ -105,8 +103,8 @@ describe("deleteWorktree", () => {
       strictEqual(result.value.changedFiles, undefined);
     }
 
-    strictEqual(validateMock.mock.calls.length, 1);
-    deepStrictEqual(validateMock.mock.calls[0].arguments, [
+    strictEqual(validateWorktreeExistsMock.mock.calls.length, 1);
+    deepStrictEqual(validateWorktreeExistsMock.mock.calls[0].arguments, [
       "/test/repo",
       "feature",
     ]);
@@ -114,35 +112,28 @@ describe("deleteWorktree", () => {
     strictEqual(executeGitCommandInDirectoryMock.mock.calls.length, 1);
     deepStrictEqual(executeGitCommandInDirectoryMock.mock.calls[0].arguments, [
       "/test/repo/.git/phantom/worktrees/feature",
-      "status --porcelain",
+      ["status", "--porcelain"],
     ]);
 
     strictEqual(executeGitCommandMock.mock.calls.length, 2);
     deepStrictEqual(executeGitCommandMock.mock.calls[0].arguments, [
-      'worktree remove "/test/repo/.git/phantom/worktrees/feature"',
+      ["worktree", "remove", "/test/repo/.git/phantom/worktrees/feature"],
       { cwd: "/test/repo" },
     ]);
     deepStrictEqual(executeGitCommandMock.mock.calls[1].arguments, [
-      'branch -D "feature"',
+      ["branch", "-D", "feature"],
       { cwd: "/test/repo" },
     ]);
   });
 
   it("should fail when worktree does not exist", async () => {
-    const validateMock = mock.fn(() =>
+    resetMocks();
+    validateWorktreeExistsMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: false,
         message: "Worktree 'nonexistent' does not exist",
       }),
     );
-
-    mock.module("./validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    const { deleteWorktree } = await import("./delete.ts");
 
     const result = await deleteWorktree("/test/repo", "nonexistent");
 
@@ -154,34 +145,20 @@ describe("deleteWorktree", () => {
   });
 
   it("should fail when uncommitted changes exist without force", async () => {
-    const validateMock = mock.fn(() =>
+    resetMocks();
+    validateWorktreeExistsMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
 
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({
         stdout: "M file1.txt\nA file2.txt\n?? file3.txt",
         stderr: "",
       }),
     );
-
-    mock.module("./validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: mock.fn(),
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { deleteWorktree } = await import("./delete.ts");
 
     const result = await deleteWorktree("/test/repo", "feature");
 
@@ -196,43 +173,29 @@ describe("deleteWorktree", () => {
   });
 
   it("should delete worktree with uncommitted changes when force is true", async () => {
-    const validateMock = mock.fn(() =>
+    resetMocks();
+    validateWorktreeExistsMock.mock.mockImplementation(() =>
       Promise.resolve({
         exists: true,
         path: "/test/repo/.git/phantom/worktrees/feature",
       }),
     );
 
-    const executeGitCommandMock = mock.fn((command: string) => {
-      if (command.includes("worktree remove")) {
+    executeGitCommandMock.mock.mockImplementation((command) => {
+      if (command[0] === "worktree" && command[1] === "remove") {
         return Promise.resolve({ stdout: "", stderr: "" });
       }
-      if (command.includes("branch -D")) {
+      if (command[0] === "branch" && command[1] === "-D") {
         return Promise.resolve({ stdout: "", stderr: "" });
       }
       return Promise.reject(new Error("Unexpected command"));
     });
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({
         stdout: "M file1.txt\nA file2.txt",
         stderr: "",
       }),
     );
-
-    mock.module("./validate.ts", {
-      namedExports: {
-        validateWorktreeExists: validateMock,
-      },
-    });
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { deleteWorktree } = await import("./delete.ts");
 
     const result = await deleteWorktree("/test/repo", "feature", {
       force: true,
@@ -251,19 +214,15 @@ describe("deleteWorktree", () => {
 });
 
 describe("getWorktreeStatus", () => {
+  const resetMocks = () => {
+    executeGitCommandInDirectoryMock.mock.resetCalls();
+  };
+
   it("should return no uncommitted changes when git status is clean", async () => {
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({ stdout: "", stderr: "" }),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: mock.fn(),
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { getWorktreeStatus } = await import("./delete.ts");
 
     const status = await getWorktreeStatus("/test/worktree");
 
@@ -273,26 +232,18 @@ describe("getWorktreeStatus", () => {
     strictEqual(executeGitCommandInDirectoryMock.mock.calls.length, 1);
     deepStrictEqual(executeGitCommandInDirectoryMock.mock.calls[0].arguments, [
       "/test/worktree",
-      "status --porcelain",
+      ["status", "--porcelain"],
     ]);
   });
 
   it("should return uncommitted changes when git status shows changes", async () => {
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.resolve({
         stdout: "M file1.txt\nA file2.txt\n?? file3.txt",
         stderr: "",
       }),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: mock.fn(),
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { getWorktreeStatus } = await import("./delete.ts");
 
     const status = await getWorktreeStatus("/test/worktree");
 
@@ -301,18 +252,10 @@ describe("getWorktreeStatus", () => {
   });
 
   it("should return no changes when git status fails", async () => {
-    const executeGitCommandInDirectoryMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandInDirectoryMock.mock.mockImplementation(() =>
       Promise.reject(new Error("Not a git repository")),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: mock.fn(),
-        executeGitCommandInDirectory: executeGitCommandInDirectoryMock,
-      },
-    });
-
-    const { getWorktreeStatus } = await import("./delete.ts");
 
     const status = await getWorktreeStatus("/test/worktree");
 
@@ -322,19 +265,15 @@ describe("getWorktreeStatus", () => {
 });
 
 describe("removeWorktree", () => {
+  const resetMocks = () => {
+    executeGitCommandMock.mock.resetCalls();
+  };
+
   it("should remove worktree successfully", async () => {
-    const executeGitCommandMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandMock.mock.mockImplementation(() =>
       Promise.resolve({ stdout: "", stderr: "" }),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: mock.fn(),
-      },
-    });
-
-    const { removeWorktree } = await import("./delete.ts");
 
     await removeWorktree(
       "/test/repo",
@@ -343,29 +282,21 @@ describe("removeWorktree", () => {
 
     strictEqual(executeGitCommandMock.mock.calls.length, 1);
     deepStrictEqual(executeGitCommandMock.mock.calls[0].arguments, [
-      'worktree remove "/test/repo/.git/phantom/worktrees/feature"',
+      ["worktree", "remove", "/test/repo/.git/phantom/worktrees/feature"],
       { cwd: "/test/repo" },
     ]);
   });
 
   it("should use force removal when regular removal fails", async () => {
+    resetMocks();
     let callCount = 0;
-    const executeGitCommandMock = mock.fn(() => {
+    executeGitCommandMock.mock.mockImplementation(() => {
       callCount++;
       if (callCount === 1) {
         return Promise.reject(new Error("Worktree is dirty"));
       }
       return Promise.resolve({ stdout: "", stderr: "" });
     });
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: mock.fn(),
-      },
-    });
-
-    const { removeWorktree } = await import("./delete.ts");
 
     await removeWorktree(
       "/test/repo",
@@ -374,28 +305,25 @@ describe("removeWorktree", () => {
 
     strictEqual(executeGitCommandMock.mock.calls.length, 2);
     deepStrictEqual(executeGitCommandMock.mock.calls[0].arguments, [
-      'worktree remove "/test/repo/.git/phantom/worktrees/feature"',
+      ["worktree", "remove", "/test/repo/.git/phantom/worktrees/feature"],
       { cwd: "/test/repo" },
     ]);
     deepStrictEqual(executeGitCommandMock.mock.calls[1].arguments, [
-      'worktree remove --force "/test/repo/.git/phantom/worktrees/feature"',
+      [
+        "worktree",
+        "remove",
+        "--force",
+        "/test/repo/.git/phantom/worktrees/feature",
+      ],
       { cwd: "/test/repo" },
     ]);
   });
 
   it("should throw error when both regular and force removal fail", async () => {
-    const executeGitCommandMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandMock.mock.mockImplementation(() =>
       Promise.reject(new Error("Permission denied")),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: mock.fn(),
-      },
-    });
-
-    const { removeWorktree } = await import("./delete.ts");
 
     try {
       await removeWorktree(
@@ -404,7 +332,7 @@ describe("removeWorktree", () => {
       );
       throw new Error("Expected removeWorktree to throw");
     } catch (error) {
-      strictEqual((error as Error).message, "Failed to remove worktree");
+      strictEqual(error.message, "Failed to remove worktree");
     }
 
     strictEqual(executeGitCommandMock.mock.calls.length, 2);
@@ -412,19 +340,15 @@ describe("removeWorktree", () => {
 });
 
 describe("deleteBranch", () => {
+  const resetMocks = () => {
+    executeGitCommandMock.mock.resetCalls();
+  };
+
   it("should delete branch successfully", async () => {
-    const executeGitCommandMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandMock.mock.mockImplementation(() =>
       Promise.resolve({ stdout: "", stderr: "" }),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: mock.fn(),
-      },
-    });
-
-    const { deleteBranch } = await import("./delete.ts");
 
     const result = await deleteBranch("/test/repo", "feature");
 
@@ -434,24 +358,16 @@ describe("deleteBranch", () => {
     }
     strictEqual(executeGitCommandMock.mock.calls.length, 1);
     deepStrictEqual(executeGitCommandMock.mock.calls[0].arguments, [
-      'branch -D "feature"',
+      ["branch", "-D", "feature"],
       { cwd: "/test/repo" },
     ]);
   });
 
   it("should return error when branch deletion fails", async () => {
-    const executeGitCommandMock = mock.fn(() =>
+    resetMocks();
+    executeGitCommandMock.mock.mockImplementation(() =>
       Promise.reject(new Error("Branch not found")),
     );
-
-    mock.module("../git/executor.ts", {
-      namedExports: {
-        executeGitCommand: executeGitCommandMock,
-        executeGitCommandInDirectory: mock.fn(),
-      },
-    });
-
-    const { deleteBranch } = await import("./delete.ts");
 
     const result = await deleteBranch("/test/repo", "feature");
